==> Chunk8-Data Analysis/7-main_figure/01-fig2_fig3.R <==
library(stringr)
library(purrr)
library(dplyr)
library(tidyr)
library(ggplot2)
library(patchwork)
library(funkyheatmap)


################################################################################
#######################   Figure 2 --- detail heatmap  #########################
################################################################################
source("Chunk8-Data Analysis/7-main_figure/04-utils.R")
accuracy <- readRDS("Chunk8-Data Analysis/2-accuracy/accuracy_long_data.rds")
spatial_accuracy <- readRDS("Chunk8-Data Analysis/2-accuracy/spatial_accuracy_data.rds")
functionality <- readRDS("Chunk8-Data Analysis/3-functionality/functionality_long_data.rds")
scalability <- readRDS("Chunk8-Data Analysis/4-scalability/score_data.rds")
usability <- readRDS("Chunk8-Data Analysis/5-usability/usability.rds")
method <- openxlsx::read.xlsx("Chunk1-Data preparation/methods.xlsx", sheet = 1, colNames = TRUE) %>% 
  select(-2)
colnames(method) <- c("Method",
                      "Category",
                      "Language",
                      "Model",
                      "Model Category",
                      "SRT/Omics Simulator",
                      "Prior Information",
                      "Simulate Groups",
                      "Simulate DEGs",
                      "Simulate Batches",
                      "Simulate Trajectory")
method_names <- method$Method
data_info <- openxlsx::read.xlsx("./Chunk1-Data preparation/evaluation_datasets.xlsx", rowNames = TRUE, sep.names = " ") %>% 
  mutate(
    Data = str_split(`Dataset ID`, "_", simplify = TRUE)[, 1]
  ) %>% 
  select(Data, Platform, `Quantification Strategy`)

accuracy <- accuracy %>% 
  full_join(data_info, by = "Data") %>% 
  relocate(Platform, .after = "Data") %>% 
  mutate(
    Platform = case_when(
      Platform == "Smart-seq2\r\n10X Genomics" ~ "Mix sources1",
      Platform == "CEL-seq\r\nCEL-seq2" ~ "Mix sources2",
      TRUE ~ Platform
    ),
    Type = case_when(
      Data %in% paste0("data", 1:101) ~ "scRNA-seq data",
      Data %in% paste0("data", 101:152) ~ "spatial transcriptome data"
    )
  ) %>% 
  relocate(Type, .after = "Platform")

spatial_accuracy <- spatial_accuracy %>% 
  full_join(data_info, by = "Data") %>% 
  relocate(Platform, .after = "Data") %>% 
  mutate(
    Platform = case_when(
      Platform == "Smart-seq2\r\n10X Genomics" ~ "Mix sources1",
      Platform == "CEL-seq\r\nCEL-seq2" ~ "Mix sources2",
      TRUE ~ Platform
    ),
    Type = case_when(
      Data %in% paste0("data", 1:101) ~ "scRNA-seq data",
      Data %in% paste0("data", 101:152) ~ "spatial transcriptome data"
    )
  ) %>% 
  relocate(Type, .after = "Platform")

accuracy_plot <- accuracy_process_function(accuracy)
spatial_accuracy_plot <- accuracy_process_function(spatial_accuracy)
spatial_accuracy_plot <- spatial_accuracy_plot %>% 
  select(c(1:4, 6:9, 11:21)) %>% 
  slice(-42)
saveRDS(spatial_accuracy_plot, file = "Chunk8-Data Analysis/2-accuracy/spatial_accuracy_plot_data.rds")

############################# errors
errors <- openxlsx::read.xlsx(xlsxFile = "./Chunk8-Data Analysis/6-error_reason/error.xlsx", sheet = 1)
errors <- errors %>% 
  mutate(
    method = str_split(error_file, "_", simplify = TRUE)[, 1],
    data = case_when(
      str_starts(str_split(error_file, "_", simplify = TRUE)[, 2], "data") ~ str_split(error_file, "_", simplify = TRUE)[, 2],
      TRUE ~ str_split(error_file, "_", simplify = TRUE)[, 3]
    ),
    data_type = case_when(
      data %in% paste0("data", 1:101) ~ "single-cell data",
      data %in% paste0("data", 102:152) ~ "spatial data"
    )
  )

### error per method, data_type and error category
method_error <- errors %>% 
  group_by(method, category, data_type) %>% 
  summarise(
    count = n()
  )
saveRDS(method_error, file = "./Chunk8-Data Analysis/6-error_reason/method_error.rds")

### proportion of errors
simulated_list <- list.files("/Volumes/Elements/sim_bench/simulation_data/")
succeed_methods <- table(str_split(simulated_list, pattern = "_", simplify = TRUE)[, 1]) %>% as.data.frame()
colnames(succeed_methods) <- c("Method", "Succeed")
method_error_count <- method_error %>% 
  group_by(method) %>% 
  summarise(
    Failed = sum(count)
  )
colnames(method_error_count)[1] <- "Method"
methods_error_table <- full_join(succeed_methods, method_error_count, by = "Method") %>% 
  mutate(
    across(c(2,3), ~ replace_na(.x, 0)),
    Total = Succeed + Failed,
    error_proportion = Failed / Total,
    error = paste0(as.character(round(error_proportion, digits = 3) * 100), "%")
  )

### proportion of errors per category
error_proportion_category <- method_error %>% 
  group_by(method) %>% 
  reframe(
    error_proportion_category = count/sum(count),
    category = category
  )
category <- unique(error_proportion_category$category)
error_proportion_category <- map(methods_error_table$Method, .f = function(x){
  values <- error_proportion_category %>% 
    filter(method == x) %>% 
    pull(error_proportion_category)
  if(S4Vectors::isEmpty(values)){
    values <- rep(0, 11)
    names(values) <- category
  }else{
    names(values) <- error_proportion_category %>% 
      filter(method == x) %>% 
      pull(category)
  }
  values
})

error_proportion_category_tibble <- map_dfr(1:length(error_proportion_category), .f = function(x){
  tmp <- error_proportion_category[[x]][category]
  names(tmp) <- category
  tmp[is.na(tmp)] <- 0
  error_proportion_category[[x]] <- tmp
  a <- tibble(
    Method = methods_error_table$Method[x],
    Reason = error_proportion_category[x]
  )
  a
})

error_info <- full_join(methods_error_table, error_proportion_category_tibble, by = "Method")
saveRDS(error_info, file = "./Chunk8-Data Analysis/6-error_reason/error_info.rds")

### Add to accuracy
accuracy_plot <- accuracy_plot %>% 
  full_join(., error_info, by = "Method")


### summarize functionality score per metric
functionality <- functionality %>% 
  full_join(data_info, by = "Data") %>% 
  relocate(Platform, .after = "Data") %>% 
  mutate(
    Platform = case_when(
      Platform == "Smart-seq2\r\n10X Genomics" ~ "Mix sources1",
      Platform == "CEL-seq\r\nCEL-seq2" ~ "Mix sources2",
      TRUE ~ Platform
    ),
    Type = case_when(
      Data %in% paste0("data", 1:101) ~ "scRNA-seq data",
      Data %in% paste0("data", 101:152) ~ "spatial transcriptome data"
    )
  ) %>% 
  relocate(Type, .after = "Platform")

functionality_plot <- functionality_process_function(functionality)

### scalability score
colnames(scalability) <- c("Method",
                           "estimation time",
                           "estimation memory",
                           "simulation time",
                           "simulation memory",
                           "time",
                           "memory",
                           "scalability",
                           "Cor(time_estimation)",
                           "Cor(memory_estimation)",
                           "Cor(time_simulation)",
                           "Cor(memory_simulation)")
scalability_plot <- readRDS("./Chunk8-Data Analysis/4-scalability/scalability_plot_data.rds")
colnames(scalability_plot)[1] <- "Method"
scalability <- scalability %>% 
  full_join(scalability_plot, by = "Method")
scalability$`Cor(time_estimation)` <- as.character(scalability$`Cor(time_estimation)`)
scalability$`Cor(memory_estimation)` <- as.character(scalability$`Cor(memory_estimation)`)
scalability$`Cor(time_simulation)` <- as.character(scalability$`Cor(time_simulation)`)
scalability$`Cor(memory_simulation)` <- as.character(scalability$`Cor(memory_simulation)`)

### usability
colnames(usability)[1] <- "Method"


### overall data
overall_data <- method %>% 
  left_join(., accuracy_plot, by = "Method") %>% 
  left_join(., functionality_plot, by = "Method") %>% 
  left_join(., scalability, by = "Method") %>% 
  left_join(., usability, by = "Method") %>% 
  relocate("accuracy", .after = `Simulate Trajectory`) %>% 
  relocate("functionality", .after = accuracy) %>% 
  relocate("scalability", .after = functionality) %>% 
  relocate("usability", .after = scalability)

overall_data <- overall_data %>% 
  mutate(
    overall = apply(overall_data %>% select(12:15), MARGIN = 1, function(x){mean(c(x[1], x[2], x[3], x[4]), na.rm = TRUE)})
  ) %>% 
  relocate(overall, .after = `Simulate Trajectory`)


overall_data <- overall_data %>% 
  mutate(
    id = Method,
  ) %>% 
  relocate(id, .before = Method) %>% 
  select(-2)

overall_data <- overall_data %>%
  mutate(
    across(all_of(c("Model",
                    "Model Category",
                    "Prior Information",
                    "Simulate Groups",
                    "Simulate DEGs",
                    "Simulate Batches",
                    "Simulate Trajectory")), ~ replace_na(.x, ""))
  ) %>% 
  as_tibble()

### arrange method
arrange_by_group <- function(tibble){
  groups <- unique(tibble %>% pull("Category"))
  result <- tibble()
  for(i in groups){
    tmp <- tibble %>% 
      filter(Category == i) %>% 
      arrange(desc(overall))
    result <- rbind(result, tmp)
  }
  return(result)
}

overall_data <- arrange_by_group(tibble = overall_data)
saveRDS(overall_data, file = "./Chunk8-Data Analysis/overall_data.rds")
###--------------------------------------------------------------------------###
###                            Fig2. Summary
###--------------------------------------------------------------------------###

### define column information
column_info <- tribble(
  ~ id,                    ~group,             ~name,                     ~geom,       ~palette,    ~options,
  "id",                    "method",           "",                        "text",      NA,          list(hjust = 0, width = 3),
  "Language",              "method_info",      "Language",                "text",      NA,          list(width = 1.5),
  "Model",                 "method_info",      "Model",                   "text",      NA,          list(width = 6),
  "SRT/Omics Simulator",   "method_info",      "SRT/Omics Simulator",     "text",      NA,          list(width = 0.5),
  "Prior Information",     "method_info",      "Prior Information",       "text",      NA,          list(width = 6),
  "Simulate Groups",       "function",         "Groups",                  "text",      NA,          list(width = 1),
  "Simulate DEGs",         "function",         "DEGs",                    "text",      NA,          list(width = 1),
  "Simulate Batches",      "function",         "Batches",                 "text",      NA,          list(width = 1),
  "Simulate Trajectory",   "function",         "Trajectory",              "text",      NA,          list(width = 1),
  "functionality_score",   "overall",          "Functionality Score",     "text",      NA,          list(width = 1.2),
  "functionality",         "overall",          NA,                        "bar",       "palette3",  list(width = 5),
  "accuracy_score",        "overall",          "Accuracy Score",          "text",      NA,          list(width = 1.2),
  "accuracy",              "overall",          NA,                        "bar",       "palette2",  list(width = 5),
  "scalability_score",     "overall",          "Scalability Score",       "text",      NA,          list(width = 1.2),
  "scalability",           "overall",          NA,                        "bar",       "palette5",  list(width = 5),
  "usability_score",       "overall",          "Usability Score",         "text",      NA,          list(width = 1.2),
  "usability",             "overall",          NA,                        "bar",       "palette4",  list(width = 5),
  "overall_score",         "overall",          "Overall Score",           "text",      NA,          list(width = 1.2),
  "overall",               "overall",          NA,                        "bar",       "palette1",  list(width = 4.2),
)

### column grouping

column_groups <- tribble(
  ~ Experiment,                ~Category,                                        ~group,         ~palette,
  "Method Characteristics",    "",                                               "method",       "palette1",
  "Method Characteristics",    "",                                               "method_info",  "palette1",
  "Method Characteristics",    "Functionality",                                  "function",     "palette1",
  "Evaluation Summary",        "Scores of overall performace and four criteria", "overall",      "palette1"
)

### row info
row_info <- tribble(
   ~group,         ~ id,                   
   "Class 1",      "PROSSTT",
   "Class 1",      "Splat-paths",
   "Class 1",      "scDesign3-tree",
   "Class 1",      "phenopath",
   "Class 1",      "SCRIP-paths",
   "Class 1",      "ESCO-tree",
   "Class 1",      "SplatPop-paths",
   "Class 1",      "ESCO-traj",
   "Class 1",      "MFA",
   "Class 1",      "dyngen",
   "Class 1",      "scMultiSim-tree",
   "Class 1",      "SymSim",
   "Class 1",      "TedSim",
   "Class 1",      "dyntoy",
   "Class 1",      "VeloSim",
   "Class 2",      "Splat",
   "Class 2",      "SPARSim",
   "Class 2",      "SCRIP-BP",
   "Class 2",      "SCRIP-GP-commonBCV",
   "Class 2",      "SCRIP-GP-trendedBCV",
   "Class 2",      "SplatPop",
   "Class 2",      "SCRIP-BGP-commonBCV",
   "Class 2",      "SCRIP-BGP-trendedBCV",
   "Class 2",      "powsimR",
   "Class 2",      "SPsimSeq",
   "Class 3",      "Lun",
   "Class 3",      "muscat",
   "Class 3",      "scDesign",
   "Class 3",      "scDD",
   "Class 3",      "ESCO",
   "Class 3",      "scDesign3",
   "Class 3",      "scMultiSim",
   "Class 3",      "zingeR",
   "Class 3",      "zinbwaveZinger",
   "Class 4",      "SRTsim",
   "Class 4",      "Lun2",
   "Class 4",      "scDesign2",
   "Class 4",      "POWSC",
   "Class 4",      "BASiCS",
   "Class 4",      "hierarchicell",
   "Class 4",      "SparseDC",
   "Class 4",      "scGAN",
   "Class 4",      "SimBPDD",
   "Class 5",      "Simple",
   "Class 5",      "Kersplat",
   "Class 5",      "zinbwave",
   "Class 5",      "dropsim",
   "Class 5",      "BEARscc",
   "Class 5",      "CancerInSilico"
)


### row groups
row_groups <- tribble(
  ~group,         ~ id,                   
  "Class 1",      "Class 1 Method",
  "Class 2",      "Class 2 Method",
  "Class 3",      "Class 3 Method",
  "Class 4",      "Class 4 Method",
  "Class 5",      "Class 5 Method"
)

### determine palettes
error_reasons_palettes <- RColorBrewer::brewer.pal(12, "Set3")[-c(9,10)]
a <- overall_data[1, ] %>% pull("Reason")
category <- names(a[[1]])
rm(a)
names(error_reasons_palettes) <- category

palettes <- tribble(
  ~palette,             ~colours,
  "palette1",           grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "Greys")[-1]))(101),
  "palette2",           grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "Blues")))(101),
  "palette3",           grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "Reds")[-9]))(101),
  "palette4",           grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "YlOrBr")[-9]))(101),
  "palette5",           grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "Greens")[-9]))(101),
  "error_reasons",      error_reasons_palettes,
  "scala_rect",         grDevices::colorRampPalette(rev(RColorBrewer::brewer.pal(9, "Reds")[-9]))(101),
  "white6black4",       c("white","white","white","black","black","black","black","black","black","black")
)

summary_plot <- funky_heatmap(data = overall_data %>%
                                mutate(overall_score = as.character(round(overall, digits = 2)),
                                       accuracy_score = as.character(round(accuracy, digits = 2)),
                                       functionality_score = as.character(round(functionality, digits = 2)),
                                       scalability_score = as.character(round(scalability, digits = 2)),
                                       usability_score = as.character(round(usability, digits = 2))),
                              column_info = column_info,
                              column_groups = column_groups,
                              row_info = row_info,
                              row_groups = row_groups,
                              palettes = palettes,
                              scale_column = FALSE,
                              expand = c(xmin = 0.5, xmax = 0.5))

library(Cairo)
ggsave(summary_plot,
       filename = "/Users/duohongrui/Desktop/sim-article/figures/Fig2.pdf",
       units = "in",
       width = 17,
       height = 24,
       device = cairo_pdf)

###--------------------------------------------------------------------------###
###                             Fig.3
###--------------------------------------------------------------------------###

### define column information
column_info <- column_info <- tribble(
  ~ id,                                     ~group,             ~name,                     ~geom,       ~palette,         ~options,
  "id",                                     "method",           "",                        "text",      NA,               list(hjust = 0,
                                                                                                                               width = 4,
                                                                                                                               size = 4, 
                                                                                                                               legend = FALSE),
  "KDE",                                    "property",         "Kernel Density",          "funkyrect", "palette2",       NULL,
  "KS",                                     "property",         "KS Distance",             "funkyrect", "palette2",       NULL,
  "MAD",                                    "property",         "MAD",                     "funkyrect", "palette2",       NULL,
  "MAE",                                    "property",         "MAE",                     "funkyrect", "palette2",       NULL,
  "OV",                                     "property",         "Overlapping Index",       "funkyrect", "palette2",       NULL,
  "RMSE",                                   "property",         "RMSE",                    "funkyrect", "palette2",       NULL,
  "bhattacharyya",                          "property",         "BH Distance",             "funkyrect", "palette2",       NULL,
  "multiKS",                                "property",         "multiKS",                 "funkyrect", "palette2",       NULL,
  "Group_score",                            "group",            "Group Score",             "funkyrect", "palette3",       NULL,
  "CDI",                                    "group",            "CDI",                     "funkyrect", "palette3",       NULL,
  "ROUGE",                                  "group",            "ROUGE",                   "funkyrect", "palette3",       NULL,
  "silhouette",                             "group",            "ASW",                     "funkyrect", "palette3",       NULL,
  "dunn",                                   "group",            "Dunn Index",              "funkyrect", "palette3",       NULL,
  "connectivity",                           "group",            "Connectivity",            "funkyrect", "palette3",       NULL,
  "DB_index",                               "group",            "DB Index",                "funkyrect", "palette3",       NULL,
  "DEGs_score",                             "DEGs",             "DEGs Score",              "funkyrect", "palette3",       NULL,
  "distribution_score",                     "DEGs",             "Distribution Score",      "funkyrect", "palette3",       NULL,
  "true_proportion",                        "DEGs",             "True Proportion",         "funkyrect", "palette3",       NULL,
  "Accuracy",                               "DEGs",             "Accuracy",                "funkyrect", "palette3",       NULL,
  "Precision",                              "DEGs",             "Precision",               "funkyrect", "palette3",       NULL,
  "Recall",                                 "DEGs",             "Recall",                  "funkyrect", "palette3",       NULL,
  "F1",                                     "DEGs",             "F1 Score",                "funkyrect", "palette3",       NULL,
  "AUC",                                    "DEGs",             "AUC",                     "funkyrect", "palette3",       NULL,
  "Batch_score",                            "batch",            "Batch Score",             "funkyrect", "palette3",       NULL,
  "cms",                                    "batch",            "CMS",                     "funkyrect", "palette3",       NULL,
  "LISI",                                   "batch",            "LISI",                    "funkyrect", "palette3",       NULL,
  "mm",                                     "batch",            "Mixing Metric",           "funkyrect", "palette3",       NULL,
  "shannon_entropy",                        "batch",            "Shannon Entropy",         "funkyrect", "palette3",       NULL,
  "kBET",                                   "batch",            "kBET",                    "funkyrect", "palette3",       NULL,
  "AWS_batch",                              "batch",            "AWS batch",               "funkyrect", "palette3",       NULL,
  "pcr",                                    "batch",            "PCR",                     "funkyrect", "palette3",       NULL,
  "Trajectory_score",                       "trajectory",       "Trajectory Score",        "funkyrect", "palette3",       NULL,
  "HIM",                                    "trajectory",       "HIM",                     "funkyrect", "palette3",       NULL,
  "F1_branches",                            "trajectory",       "F1_branches",             "funkyrect", "palette3",       NULL,
  "F1_milestones",                          "trajectory",       "F1_milestones",           "funkyrect", "palette3",       NULL,
  "Cor_dist",                               "trajectory",       "Cor_dist",                "funkyrect", "palette3",       NULL,
  "estimation time",                        "scala",            "Time (Estimation)",       "funkyrect", "palette5",       NULL,
  "estimation memory",                      "scala",            "Memory (Estimation)",     "funkyrect", "palette5",       NULL,
  "simulation time",                        "scala",            "Time (Simulation)",       "funkyrect", "palette5",       NULL,
  "simulation memory",                      "scala",            "Memory (Simulation)",     "funkyrect", "palette5",       NULL,
  "Avalability",                            "usability",        "Avalability",             "funkyrect", "palette4",       NULL,
  "Code",                                   "usability",        "Code",                    "funkyrect", "palette4",       NULL,
  "Documentation",                          "usability",        "Documentation",           "funkyrect", "palette4",       NULL,
  "Evaluation",                             "usability",        "Self-evaluation",         "funkyrect", "palette4",       NULL,
  "Maintenance",                            "usability",        "Maintenance",             "funkyrect", "palette4",       NULL,
  "Paper",                                  "usability",        "Paper Quality",           "funkyrect", "palette4",       NULL,
  "error",                                  "error",            "Proportion of errors",    "text",      NA,               list(hjust = 0, width = 2, size = 4),
  "Reason",                                 "error",            "Reasons",                 "pie",       "error_reasons",  NULL
)

### column grouping

column_groups <- tribble(
  ~ Experiment,     ~Category,                                       ~group,           ~palette,
  "Method",         "",                                              "method",         "palette1",
  "Accuracy",       "Per metric",                                    "property",       "palette2",
  "Functionality",  "Group",                                         "group",          "palette3",
  "Functionality",  "DEGs",                                          "DEGs",           "palette3",
  "Functionality",  "Batch",                                         "batch",          "palette3",
  "Functionality",  "Trajectory",                                    "trajectory",     "palette3",
  "Scalability",    "Time consuming \n memory usage",                "scala",          "palette5",
  "Usability",      "Quality of the codes \nand softwares",          "usability",      "palette4",
  "Usability",      "Error",                                         "error",          "palette4",
)


fig2_plot <- funky_heatmap(data = overall_data,
                           column_info = column_info,
                           column_groups = column_groups,
                           row_info = row_info,
                           row_groups = row_groups,
                           palettes = palettes,
                           scale_column = FALSE,
                           expand = c(xmin = 0.5, xmax = 3),
                           col_annot_offset = 4)

ggsave(fig2_plot,
       filename = "/Users/duohongrui/Desktop/sim-article/figures/Fig3.pdf",
       units = "in",
       width = 16,
       height = 19)


###--------------------------------------------------------------------------###
###                       Supplementary Fig.3 (Accuracy)
###--------------------------------------------------------------------------###

### define column information
column_info <- tribble(
  ~ id,                                     ~group,             ~name,                     ~geom,       ~palette,         ~options,
  "id",                                     "method",           "",                        "text",      NA,               list(hjust = 0,
                                                                                                                               width = 4,
                                                                                                                               size = 4,
                                                                                                                               legend = FALSE),
  "KDE",                                    "metric",           "Kernel Density",          "funkyrect", "palette2",       NULL,
  "KS",                                     "metric",           "KS Distance",             "funkyrect", "palette2",       NULL,
  "MAD",                                    "metric",           "MAD",                     "funkyrect", "palette2",       NULL,
  "MAE",                                    "metric",           "MAE",                     "funkyrect", "palette2",       NULL,
  "OV",                                     "metric",           "Overlapping Index",       "funkyrect", "palette2",       NULL,
  "RMSE",                                   "metric",           "RMSE",                    "funkyrect", "palette2",       NULL,
  "bhattacharyya",                          "metric",           "BH Distance",             "funkyrect", "palette2",       NULL,
  "multiKS",                                "metric",           "multiKS",                 "funkyrect", "palette2",       NULL,
  "LS",                                     "property",         "LS",                      "funkyrect", "palette2",       NULL,
  "FZC",                                    "property",         "FZC",                     "funkyrect", "palette2",       NULL,
  "CCC",                                    "property",         "CCC",                     "funkyrect", "palette2",       NULL,
  "TMM",                                    "property",         "TMM",                     "funkyrect", "palette2",       NULL,
  "ELS",                                    "property",         "ELS",                     "funkyrect", "palette2",       NULL,
  "FCO",                                    "property",         "FCO",                     "funkyrect", "palette2",       NULL,
  "RLZ",                                    "property",         "RLZ",                     "funkyrect", "palette2",       NULL,
  "ME",                                     "property",         "ME",                      "funkyrect", "palette2",       NULL,
  "SD",                                     "property",         "SD",                      "funkyrect", "palette2",       NULL,
  "CV",                                     "property",         "CV",                      "funkyrect", "palette2",       NULL,
  "FZG",                                    "property",         "FZG",                     "funkyrect", "palette2",       NULL,
  "FGO",                                    "property",         "FGO",                     "funkyrect", "palette2",       NULL,
  "RMS",                                    "property",         "RMS",                     "funkyrect", "palette2",       NULL,
  "RMZ",                                    "property",         "RMZ",                     "funkyrect", "palette2",       NULL,
  "RDM",                                    "property",         "RDM",                     "funkyrect", "palette2",       NULL,
  "acc_MARS-seq",                           "platform",         "MARS-Seq",                "funkyrect", "palette2",       NULL,
  "acc_10X Genomics",                       "platform",         "10X Genomics",            "funkyrect", "palette2",       NULL,
  "acc_Smart-seq2",                         "platform",         "Smart-seq2",              "funkyrect", "palette2",       NULL,
  "acc_Mix sources1",                       "platform",         "Mix sources1",            "funkyrect", "palette2",       NULL,
  "acc_CEL-seq",                            "platform",         "CEL-seq",                 "funkyrect", "palette2",       NULL,
  "acc_Fluidigm C1",                        "platform",         "Fluidigm C1",             "funkyrect", "palette2",       NULL,
  "acc_CEL-seq2",                           "platform",         "CEL-seq2",                "funkyrect", "palette2",       NULL,
  "acc_Mix sources2",                       "platform",         "Mix sources2",            "funkyrect", "palette2",       NULL,
  "acc_Drop-seq",                           "platform",         "Drop-seq",                "funkyrect", "palette2",       NULL,
  "acc_inDrop",                             "platform",         "inDrop",                  "funkyrect", "palette2",       NULL,
  "acc_STRT-seq",                           "platform",         "STRT-seq",                "funkyrect", "palette2",       NULL,
  "acc_Microwell-seq",                      "platform",         "Microwell-seq",           "funkyrect", "palette2",       NULL,
  "acc_Smart-seq",                          "platform",         "Smart-seq",               "funkyrect", "palette2",       NULL,
  "acc_ST",                                 "platform",         "ST",                      "funkyrect", "palette2",       NULL,
  "acc_HDST",                               "platform",         "HDST",                    "funkyrect", "palette2",       NULL,
  "acc_10X Visium",                         "platform",         "10X Visium",              "funkyrect", "palette2",       NULL,
  "acc_Slide-seq",                          "platform",         "Slide-Seq",               "funkyrect", "palette2",       NULL,
  "acc_Slide-seqV2",                        "platform",         "Slide-SeqV2",             "funkyrect", "palette2",       NULL,
  "acc_seqFISH",                            "platform",         "seqFISH",                 "funkyrect", "palette2",       NULL,
  "acc_seqFISH+",                           "platform",         "seqFISH+",                "funkyrect", "palette2",       NULL,
  "acc_osmFISH",                            "platform",         "osmFISH",                 "funkyrect", "palette2",       NULL,
  "acc_sci-Space",                          "platform",         "sci-Space",               "funkyrect", "palette2",       NULL,
  "acc_MERFISH",                            "platform",         "MERFISH",                 "funkyrect", "palette2",       NULL,
  "acc_Stereo-seq",                         "platform",         "Stereo-Seq",              "funkyrect", "palette2",       NULL,
  "acc_scRNA-seq data",                     "technique",        "scRNA-seq",               "funkyrect", "palette2",       NULL,
  "acc_spatial transcriptome data",         "technique",        "ST technology",           "funkyrect", "palette2",       NULL,
  "acc_read-based",                         "quantify",         "read-based",              "funkyrect", "palette2",       NULL,
  "acc_UMI-based",                          "quantify",         "UMI-based",               "funkyrect", "palette2",       NULL
)

### column grouping

column_groups <- tribble(
  ~ Experiment,     ~Category,                                       ~group,           ~palette,
  "Method",         "",                                              "method",         "palette1",
  "Accuracy",       "Per metric",                                    "metric",         "palette2",
  "Accuracy",       "Per property",                                  "property",       "palette2",
  "Accuracy",       "Per platform",                                  "platform",       "palette2",
  "Accuracy",       "Per technique",                                 "technique",      "palette2",
  "Accuracy",       "Quantification",                                "quantify",       "palette2"
)


accuracy_sup_plot <- funky_heatmap(data = overall_data,
                                   column_info = column_info,
                                   column_groups = column_groups,
                                   row_info = row_info,
                                   row_groups = row_groups,
                                   palettes = palettes,
                                   scale_column = FALSE,
                                   expand = c(xmin = 1, xmax = 4, ymin = 0.5, ymax = 0.5),
                                   col_annot_offset = 4)

ggsave(accuracy_sup_plot,
       filename = "/Users/duohongrui/Desktop/sim-article/figures/Supp_Fig_3.pdf",
       units = "in",
       width = 16,
       height = 19)


###--------------------------------------------------------------------------###
###                       Supplementary Fig.7 (Functionality)
###--------------------------------------------------------------------------###

### define column information
column_info <- column_info <- tribble(
  ~ id,                                     ~group,             ~name,                     ~geom,       ~palette,         ~options,
  "id",                                     "method",           "",                        "text",      NA,               list(hjust = 0,
                                                                                                                               width = 4,
                                                                                                                               size = 4,
                                                                                                                               legend = FALSE),
  "Group_score",                            "group",            "Group Score",             "funkyrect", "palette3",       NULL,
  "CDI",                                    "group",            "CDI",                     "funkyrect", "palette3",       NULL,
  "ROUGE",                                  "group",            "ROUGE",                   "funkyrect", "palette3",       NULL,
  "silhouette",                             "group",            "ASW",                     "funkyrect", "palette3",       NULL,
  "dunn",                                   "group",            "Dunn Index",              "funkyrect", "palette3",       NULL,
  "connectivity",                           "group",            "Connectivity",            "funkyrect", "palette3",       NULL,
  "DB_index",                               "group",            "DB Index",                "funkyrect", "palette3",       NULL,
  "DEGs_score",                             "DEGs",             "DEGs Score",              "funkyrect", "palette3",       NULL,
  "distribution_score",                     "DEGs",             "Distribution Score",      "funkyrect", "palette3",       NULL,
  "true_proportion",                        "DEGs",             "True Proportion",         "funkyrect", "palette3",       NULL,
  "Accuracy",                               "DEGs",             "Accuracy",                "funkyrect", "palette3",       NULL,
  "Precision",                              "DEGs",             "Precision",               "funkyrect", "palette3",       NULL,
  "Recall",                                 "DEGs",             "Recall",                  "funkyrect", "palette3",       NULL,
  "F1",                                     "DEGs",             "F1 Score",                "funkyrect", "palette3",       NULL,
  "AUC",                                    "DEGs",             "AUC",                     "funkyrect", "palette3",       NULL,
  "Batch_score",                            "batch",            "Batch Score",             "funkyrect", "palette3",       NULL,
  "cms",                                    "batch",            "CMS",                     "funkyrect", "palette3",       NULL,
  "LISI",                                   "batch",            "LISI",                    "funkyrect", "palette3",       NULL,
  "mm",                                     "batch",            "Mixing Metric",           "funkyrect", "palette3",       NULL,
  "shannon_entropy",                        "batch",            "Shannon Entropy",         "funkyrect", "palette3",       NULL,
  "kBET",                                   "batch",            "kBET",                    "funkyrect", "palette3",       NULL,
  "AWS_batch",                              "batch",            "AWS batch",               "funkyrect", "palette3",       NULL,
  "pcr",                                    "batch",            "PCR",                     "funkyrect", "palette3",       NULL,
  "Trajectory_score",                       "trajectory",       "Trajectory Score",        "funkyrect", "palette3",       NULL,
  "HIM",                                    "trajectory",       "HIM",                     "funkyrect", "palette3",       NULL,
  "F1_branches",                            "trajectory",       "F1_branches",             "funkyrect", "palette3",       NULL,
  "F1_milestones",                          "trajectory",       "F1_milestones",           "funkyrect", "palette3",       NULL,
  "Cor_dist",                               "trajectory",       "Cor_dist",                "funkyrect", "palette3",       NULL,
  "func_MARS-seq",                          "platform",         "MARS-Seq",                "funkyrect", "palette3",       NULL,
  "func_10X Genomics",                      "platform",         "10X Genomics",            "funkyrect", "palette3",       NULL,
  "func_Smart-seq2",                        "platform",         "Smart-seq2",              "funkyrect", "palette3",       NULL,
  "func_Mix sources1",                      "platform",         "Mix sources1",            "funkyrect", "palette3",       NULL,
  "func_CEL-seq",                           "platform",         "CEL-seq",                 "funkyrect", "palette3",       NULL,
  "func_Fluidigm C1",                       "platform",         "Fluidigm C1",             "funkyrect", "palette3",       NULL,
  "func_CEL-seq2",                          "platform",         "CEL-seq2",                "funkyrect", "palette3",       NULL,
  "func_Mix sources2",                      "platform",         "Mix sources2",            "funkyrect", "palette3",       NULL,
  "func_Drop-seq",                          "platform",         "Drop-seq",                "funkyrect", "palette3",       NULL,
  "func_inDrop",                            "platform",         "inDrop",                  "funkyrect", "palette3",       NULL,
#  "func_STRT-seq",                          "platform",         "STRT-seq",                "funkyrect", "palette2",       NULL,
#  "func_Microwell-seq",                     "platform",         "Microwell-seq",           "funkyrect", "palette3",       NULL,
  "func_Smart-seq",                         "platform",         "Smart-seq",               "funkyrect", "palette3",       NULL,
  "func_ST",                                "platform",         "ST",                      "funkyrect", "palette3",       NULL,
  "func_HDST",                              "platform",         "HDST",                    "funkyrect", "palette3",       NULL,
  "func_10X Visium",                        "platform",         "10X Visium",              "funkyrect", "palette3",       NULL,
  "func_Slide-seq",                         "platform",         "Slide-Seq",               "funkyrect", "palette3",       NULL,
  "func_Slide-seqV2",                       "platform",         "Slide-SeqV2",             "funkyrect", "palette3",       NULL,
  "func_seqFISH",                           "platform",         "seqFISH",                 "funkyrect", "palette3",       NULL,
  "func_seqFISH+",                          "platform",         "seqFISH+",                "funkyrect", "palette3",       NULL,
  "func_osmFISH",                           "platform",         "osmFISH",                 "funkyrect", "palette3",       NULL,
  "func_sci-Space",                         "platform",         "sci-Space",               "funkyrect", "palette3",       NULL,
#  "func_MERFISH",                           "platform",         "MERFISH",                 "funkyrect", "palette3",       NULL,
#  "func_Stereo-seq",                        "platform",         "Stereo-Seq",              "funkyrect", "palette3",       NULL,
  "func_scRNA-seq data",                    "technique",        "scRNA-seq",               "funkyrect", "palette3",       NULL,
  "func_spatial transcriptome data",        "technique",        "ST technology",           "funkyrect", "palette3",       NULL,
  "func_read-based",                        "quantify",         "read-based",              "funkyrect", "palette3",       NULL,
  "func_UMI-based",                         "quantify",        "UMI-based",               "funkyrect", "palette3",       NULL
)

### column grouping

column_groups <- tribble(
  ~ Experiment,     ~Category,            ~group,           ~palette,
  "Method",         "",                   "method",         "palette1",
  "Functionality",  "Group",              "group",          "palette3",
  "Functionality",  "DEGs",               "DEGs",           "palette3",
  "Functionality",  "Batch",              "batch",          "palette3",
  "Functionality",  "Trajectory",         "trajectory",     "palette3",
  "Functionality",  "Per platform",       "platform",       "palette3",
  "Functionality",  "Per technique",      "technique",      "palette3",
  "Functionality",  "Quantification",     "quantify",       "palette3"
)


functionality_sup_plot <- funky_heatmap(data = overall_data,
                                        column_info = column_info,
                                        column_groups = column_groups,
                                        row_info = row_info %>% slice(1:43),
                                        row_groups = row_groups %>% slice(1:4),
                                        palettes = palettes,
                                        scale_column = FALSE,
                                        expand = c(xmin = 0.5, xmax = 3.5),
                                        col_annot_offset = 3)

ggsave(functionality_sup_plot,
       filename = "/Users/duohongrui/Desktop/sim-article/figures/Supp_Fig_7.pdf",
       units = "in",
       width = 18,
       height = 19)
